requirements.txt
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
loguru==0.7.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-dotenv==1.0.1
scikit-learn==1.5.2
rapidfuzz==3.9.6
python-Levenshtein==0.25.1
spacy==3.7.5
sentence-transformers==3.0.1
yake==0.4.8
rake-nltk==1.0.6
networkx==3.3
pyvis==0.3.2
streamlit==1.38.0
black==24.8.0
isort==5.13.2
flake8==7.1.1
pytest==8.3.2
